refactor: move query pdf tool call handler into dedicated file
