+++ apps/hash-ai-worker-ts/src/activities/flow-activities/research-entities-action/infer-entities-from-web-page-worker-agent.ts
@@ -31,7 +31,7 @@ import { graphApiClient } from "../../shared/graph-api-client";
 import type { PermittedOpenAiModel } from "../../shared/openai-client";
 import { stringify } from "../../shared/stringify";
 import { inferEntitiesFromContentAction } from "../infer-entities-from-content-action";
-import { indexPdfFile } from "./infer-entities-from-web-page-worker-agent/llama-index/index-pdf-file";
+import { handleQueryPdfToolCall } from "./infer-entities-from-web-page-worker-agent/handle-query-pdf-tool-call";
 import type { ToolCallArguments } from "./infer-entities-from-web-page-worker-agent/tool-definitions";
 import { toolDefinitions } from "./infer-entities-from-web-page-worker-agent/tool-definitions";
 import type {
@@ -665,56 +665,9 @@ export const inferEntitiesFromWebPageWorkerAgent = async (params: {
               the initial research prompt.`,
             };
           } else if (toolCall.name === "queryPdf") {
-            const { query, fileUrl } =
-              toolCall.input as ToolCallArguments["queryPdf"];
-
-            /**
-             * Step 1: Ensure the file at the provided URL is a PDF file
-             * before fetching it, via a HEAD request.
-             */
-            const fileUrlHeadFetch = await fetch(fileUrl, { method: "HEAD" });
-
-            if (!fileUrlHeadFetch.ok) {
-              return {
-                ...toolCall,
-                output: `Failed to fetch the file at the provided URL: ${fileUrl}`,
-                isError: true,
-              };
-            }
-
-            const contentType = fileUrlHeadFetch.headers.get("Content-Type");
-
-            if (contentType && !contentType.includes("application/pdf")) {
-              return {
-                ...toolCall,
-                output: dedent(`
-                  The file at the provided URL is not a PDF file.
-                  Detected Content-Type: ${contentType}
-                `),
-                isError: true,
-              };
-            }
-
-            /**
-             * Step 2: Index the PDF file at the provided URL, and query it
-             * with the provided query.
-             */
-
-            const { vectorStoreIndex } = await indexPdfFile({ fileUrl });
-
-            const queryEngine = vectorStoreIndex.asQueryEngine();
-
-            const { response, sourceNodes } = await queryEngine.query({
-              query,
+            return await handleQueryPdfToolCall({
+              toolCall: toolCall as ParsedLlmToolCall<"queryPdf">,
             });
-
-            return {
-              ...toolCall,
-              output: dedent(`
-                Response: ${response}
-                Source Nodes: ${JSON.stringify(sourceNodes)}
-              `),
-            };
           }
 
           throw new Error(`Unimplemented tool call: ${toolCall.name}`);
